site/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "python-decouple>=3.8",
]
